% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{a1}
\alias{a1}
\alias{a2}
\alias{a3}
\title{A-sets}
\format{Data frame of x, y coordinates}
\source{
\url{http://cs.joensuu.fi/sipu/datasets/}
}
\usage{
a1

a2

a3
}
\details{
Synthetic 2-d data with varying number of vectors (N) and clusters (M). There are 150 vectors per cluster.
}
\references{
I. Kärkkäinen and P. Fränti, "Dynamic local search algorithm for the clustering problem", Research Report A-2002-6
}
\keyword{datasets}

