% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{birch1}
\alias{birch1}
\alias{birch2}
\alias{birch3}
\title{BIRCH clustering datasets.}
\format{Data frame of x, y coordinates}
\source{
\url{http://cs.joensuu.fi/sipu/datasets/}
}
\usage{
birch1

birch2

birch3
}
\description{
Clusters in regular grid structure

Clusters at a sine curve

Random sized clusters in random locations
}
\details{
Synthetic 2-d data with N=100,000 vectors and M=100 clusters
See Zhang et al., "BIRCH: A new data clustering algorithm and its applications", Data Mining and Knowledge Discovery, 1 (2), 141-182, 1997.
}
\keyword{datasets}
