% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{t48k}
\alias{t48k}
\alias{t88k}
\alias{t710k}
\alias{t58k}
\title{Chameleon Datasets}
\format{Data frame}
\usage{
t48k

t88k

t710k

t58k
}
\description{
4 sets of 2-vectors.
\itemize{
\item{"t48k"}{N=8000,M=4,D=2}
\item{"t88k"}{N=8000,M=8,D=2}
\item{"t710k"}{N=10000,M=7,D=2}
\item{"t58k"}{N=8000,M=5,D=2}
}
}
\references{
G. Karypis, E.H. Han, V. Kumar, CHAMELEON: A hierarchical clustering algorithm using dynamic modeling, IEEE Trans. on Computers, 32 (8), 68-75, 1999.
}
\keyword{datasets}
