% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringdatasets-package.r
\docType{package}
\name{clusteringdatasets}
\alias{clusteringdatasets}
\alias{clusteringdatasets-package}
\title{clusteringdatasets.}
\description{
A set of datasets useful for testing clustering algorithms.
}

