% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{dim032}
\alias{dim032}
\alias{dim064}
\alias{dim1024}
\alias{dim128}
\alias{dim256}
\alias{dim512}
\title{High-Dim Sets}
\format{Data frames}
\source{
\url{http://cs.joensuu.fi/sipu/datasets/}
}
\usage{
dim032

dim064

dim128

dim256

dim512

dim1024
}
\description{
Six sets of high-dimensional data, each with 1024 vectors and 16 gaussian clusters.
}
\references{
P. Fränti, O. Virmajoki and V. Hautamäki, "Fast agglomerative clustering using a k-nearest neighbor graph", IEEE Trans. on Pattern Analysis and Machine Intelligence, 28 (11), 1875-1881, November 2006.
}
\keyword{datasets}

