% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{kddcup04bio}
\alias{kddcup04bio}
\title{KDDCUPO4Bio}
\format{Data frame}
\source{
\url{http://cs.joensuu.fi/sipu/datasets/}
}
\usage{
kddcup04bio
}
\description{
145751 vectors, 74-D
}
\keyword{datasets}

