% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{make_blobs}
\alias{make_blobs}
\title{Make gaussian blobs}
\usage{
make_blobs(n_samples = 100, n_features = 2, centers = 3,
  cluster_std = 1, center_box = c(-10, 10), shuffle = TRUE)
}
\arguments{
\item{n_samples}{Number of points}

\item{n_features}{Dimensionality of dataset}

\item{centers}{Either the number of centers, or a matrix of the chosen centers}

\item{cluster_std}{Standard deviation of Gaussian noise. Either one number, or a vector of length equal to the number of centers}

\item{center_box}{If the centers are being generated, the bounding box within which they will be created.}

\item{shuffle}{Ignored; included for compatibility with the Python}
}
\value{
a \code{list} containining \code{samples}, a matrix of points, and \code{labels}, which identifies the cluster from which each point came.
}
\description{
Imitation of the Python \code{sklearn.datasets.make_blobs} function.
}
