% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{make_moons}
\alias{make_moons}
\title{Make two interleaving half-circles}
\usage{
make_moons(n_samples = 100, shuffle = TRUE, noise = NA)
}
\arguments{
\item{n_samples}{Number of points (will be divided equally among the circles)}

\item{shuffle}{Whether to randomize the sequence}

\item{noise}{Standard deviation of Gaussian noise applied to point positions}
}
\value{
a \code{list} containining \code{samples}, a matrix of points, and \code{labels}, which identifies the circle from which each point came.
}
\description{
Imitation of the Python \code{sklearn.datasets.make_moons} function.
}
