% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{mopsifinland}
\alias{mopsifinland}
\alias{mopsijoensu}
\title{Mopsi Data}
\format{Data frame}
\source{
\url{http://cs.uef.fi/mopsi/data/}
}
\usage{
mopsifinland

mopsijoensu
}
\description{
User locations, N = 13467 for Finland, N = 6014 for Joensuu
}
\keyword{datasets}

