% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{Circle}
\alias{Circle}
\alias{Complex1}
\alias{Complex2}
\alias{Complex3}
\alias{Complex4}
\alias{Discrete}
\alias{HiLoDensity}
\alias{JumpingRectangle}
\alias{MovingJumpingRectangle}
\alias{MovingRectangle}
\alias{Rectangle}
\alias{Ring}
\alias{RMouseRectangle}
\title{Neural Gas}
\format{Data frame}
\usage{
Circle

Complex1

Complex2

Complex3

Complex4

Discrete

HiLoDensity

JumpingRectangle

MovingJumpingRectangle

MovingRectangle

Rectangle

Ring

RMouseRectangle
}
\description{
4 sets of 2-vectors.
\itemize{
\item{"Circle"}{N=5000,D=2}
\item{"Complex1"}{N=5000,D=2}
\item{"Complex2"}{N=5000,D=2}
\item{"Complex3"}{N=5000,D=2}
\item{"Complex3"}{N=5000,D=2}
\item{"Discrete"}{N=5000,D=2}
\item{"HiLoDensity"}{N=5000,D=2}
\item{"JumpingRectangle"}{N=5000,D=2}
\item{"MovingJumpingRectangle"}{N=5000,D=2}
\item{"MovingRectangle"}{N=5000,D=2}
\item{"Rectangle"}{N=5000,D=2}
\item{"Ring"}{N=5000,D=2}
\item{"RMouseRectangle"}{N=5000,D=2}
}
}
\keyword{datasets}
