% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{cross}
\alias{cross}
\alias{d4}
\alias{face}
\alias{pie}
\alias{ring2}
\alias{sincos}
\title{Non-Convex}
\format{Data frame}
\usage{
cross

d4

face

pie

ring2

sincos
}
\description{
\itemize{
\item{"cross"}{N=2000,D=2}
\item{"d4"}{N=200,D=2}
\item{"face"}{N=500,D=2}
\item{"pie"}{N=2322,D=2}
\item{"ring2"}{N=60,D=2}
\item{"sincos"}{N=300,D=2}
}
}
\keyword{datasets}
