% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{t48k}
\alias{t48k}
\title{t48k}
\format{Data frame}
\source{
\url{http://cs.joensuu.fi/sipu/datasets/}
}
\usage{
t48k
}
\description{
8000 vectors, 2-D, forming six clusters, with noise.
}
\references{
G. Karypis, E.H. Han, V. Kumar, CHAMELEON: A hierarchical clustering algorithm using dynamic modeling, IEEE Trans. on Computers, 32 (8), 68-75, 1999.
}
\keyword{datasets}

