% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{thyroid}
\alias{breast}
\alias{glass}
\alias{thyroid}
\alias{wdbc}
\alias{wine}
\alias{yeast}
\title{UCI High Dimensional Datasets}
\format{Data frames}
\source{
\url{http://archive.ics.uci.edu/ml/}
}
\usage{
thyroid

wine

glass

yeast

breast

wdbc
}
\description{
Various high-dimensional datasets with identifying data removed
\itemize{
\item{"thyroid"}{N=215,M=2,D=5}
\item{"wine"}{N=178,M=3,D=13}
\item{"glass"}{N=214,M=7,D=9}
\item{"yeast"}{N=1484,M=10,D=8}
\item{"breast"}{N=699,M=2,D=9}
\item{"wdbc"}{N=569,M=2,D=32}
}
}
\keyword{datasets}

