% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\docType{data}
\name{thyroid}
\alias{breast}
\alias{glass}
\alias{thyroid}
\alias{wdbc}
\alias{wine}
\alias{yeast}
\title{UCI High Dimensional Datasets}
\format{Data frames}
\source{
\url{http://archive.ics.uci.edu/ml/}
}
\usage{
thyroid

wine

glass

yeast

breast

wdbc
}
\description{
Various high-dimensional datasets with identifying data removed

N = 178, M = 3, D = 13
}
\keyword{datasets}

